refactor(settings): move agent-source accessors, consts, validators to settings_agent_source.go (PRD #1021 M2)
